+++ ftlib/commlib/commlib_status.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
 
-class FrameworkStatus(Enum):
+class CommLibStatus(Enum):
     SUCCESS = 0
     FAIL = 1
