+++ ftlib/commlib/dummy_nccl/__init__.py
@@ -1,0 +1,3 @@
+__version__ = "0.0.1"
+
+from .impl import DummyNCCL  # noqa: F401
