+++ ftlib/commlib/pytorch/impl.py
@@ -1,5 +1,3 @@
-__version__ = "0.0.1"
-
 __pytorch_version__ = "1.2.0"
 
 import logging
@@ -7,17 +5,18 @@ import os
 
 import torch.distributed as dist
 
-from ..basic import BasicFramework
-from ..framework_status import FrameworkStatus
+from ftlib.commlib.basic_commlib import BasicCommLib
+from ftlib.commlib.commlib_status import CommLibStatus
 
 
-class PyTorch(BasicFramework):
+class PyTorch(BasicCommLib):
     def __init__(self, grad_sync_timeout=10, max_try=30, port=12355):
         self.type = "pytorch"
         self.grad_sync_timeout = grad_sync_timeout
         self._max_try = max_try
         self._port = port
 
+    @BasicCommLib.register_api
     def grad_sync_done(self, *args, **kwargs):
         model = None
         if "model" in kwargs.keys():
@@ -25,7 +24,7 @@ class PyTorch(BasicFramework):
         elif len(args) > 0:
             model = args[0]
         if model is None:
-            return FrameworkStatus.FAIL
+            return CommLibStatus.FAIL
         try:
             size = float(dist.get_world_size())
             for param in model.parameters():
@@ -33,9 +32,10 @@ class PyTorch(BasicFramework):
                 param.grad.data /= size
         except Exception as e:
             logging.error(str(e))
-            return FrameworkStatus.FAIL
-        return FrameworkStatus.SUCCESS
+            return CommLibStatus.FAIL
+        return CommLibStatus.SUCCESS
 
+    @BasicCommLib.register_api
     def allreduce(self, data, op="MEAN"):
         # torch.distributed.ReduceOp has no option for 'MEAN'
         # so far, we only implemented 'MEAN'
@@ -45,9 +45,11 @@ class PyTorch(BasicFramework):
             size = float(dist.get_world_size())
             data /= size
 
+    @BasicCommLib.register_api
     def broadcast(self, data, root_rank):
         dist.broadcast(data, root_rank)
 
+    @BasicCommLib.register_api
     def barrier(self):
         dist.barrier()
