+++ ftlib/consensus/consensus_status.py
@@ -5,3 +5,8 @@ class ConsensusStatus(Enum):
     SUCCESS = 0
     FAIL = 1
     SKIP_ALLREDUCE = -1
+
+
+class ConsensusMode(Enum):
+    PASSIVE = 0
+    ACTIVE = 1
